Remove the PublicKey class in favor of Key
The Key class has enough abstraction to represent both a public and a
private key. Most of the routines for PublicKey and Key classes were
the same and Key already had a class method to create representations of
public keys.

Merged all the functionality of PublicKey that was missing from Key,
updated method signatures, tests and data generation scripts.